have fixed

+++ chapter2/src/c_2_6.cc
@@ -1,17 +1,17 @@
 #include<iostream>
-int astronomical(double uns);
+double astronomical(double uns);
 int main()
 {
   using namespace std;
   double light;
   cout <<"Enter the number of light years:";
   cin>>light;
-  int Ast=astronomical(light);
+  double Ast=astronomical(light);
   cout <<light<<"light years="<<Ast<<"astronomical units"<<endl;
   return 0;
 }
 
-int astronomical(double uns)
+double astronomical(double uns)
 {
   return 63246*uns;
 }
